+++ fractal_server/app/db/__init__.py
@@ -21,6 +21,12 @@ from ...syringe import Inject
 print(__name__)
 logger = set_logger(__name__)
 
+SQLITE_WARNING_MESSAGE = (
+    "SQLite is supported (for version >=3.37) but discouraged in production. "
+    "Given its partial support for ForeignKey constraints, "
+    "database consistency cannot be guaranteed."
+)
+
 
 class DB:
     """
@@ -32,7 +38,7 @@ class DB:
         try:
             return cls._engine_async
         except AttributeError:
-            cls.set_db()
+            cls.set_async_db()
             return cls._engine_async
 
     @classmethod
@@ -40,32 +46,22 @@ class DB:
         try:
             return cls._engine_sync
         except AttributeError:
-            cls.set_db()
+            cls.set_sync_db()
             return cls._engine_sync
 
     @classmethod
-    def set_db(cls):
+    def set_async_db(cls):
         settings = Inject(get_settings)
         settings.check_db()
 
         if settings.DB_ENGINE == "sqlite":
-            logger.warning(
-                "SQLite is supported (for version >=3.37) but discouraged "
-                "in production. Given its partial support for ForeignKey "
-                "constraints, database consistency cannot be guaranteed."
-            )
-
+            logger.warning(SQLITE_WARNING_MESSAGE)
             # Set some sqlite-specific options
             engine_kwargs_async = dict(poolclass=StaticPool)
-            engine_kwargs_sync = dict(
-                poolclass=StaticPool,
-                connect_args={"check_same_thread": False},
-            )
         else:
             engine_kwargs_async = {
                 "pool_pre_ping": True,
             }
-            engine_kwargs_sync = {}
 
         cls._engine_async = create_async_engine(
             settings.DATABASE_URL,
@@ -73,18 +69,33 @@ class DB:
             future=True,
             **engine_kwargs_async,
         )
-        cls._engine_sync = create_engine(
-            settings.DATABASE_SYNC_URL,
-            echo=settings.DB_ECHO,
-            future=True,
-            **engine_kwargs_sync,
-        )
-
         cls._async_session_maker = sessionmaker(
             cls._engine_async,
             class_=AsyncSession,
             expire_on_commit=False,
             future=True,
+        )
+
+    @classmethod
+    def set_sync_db(cls):
+        settings = Inject(get_settings)
+        settings.check_db()
+
+        if settings.DB_ENGINE == "sqlite":
+            logger.warning(SQLITE_WARNING_MESSAGE)
+            # Set some sqlite-specific options
+            engine_kwargs_sync = dict(
+                poolclass=StaticPool,
+                connect_args={"check_same_thread": False},
+            )
+        else:
+            engine_kwargs_sync = {}
+
+        cls._engine_sync = create_engine(
+            settings.DATABASE_SYNC_URL,
+            echo=settings.DB_ECHO,
+            future=True,
+            **engine_kwargs_sync,
         )
 
         cls._sync_session_maker = sessionmaker(
@@ -102,14 +113,14 @@ class DB:
                 cursor.close()
 
     @classmethod
-    async def get_db(cls) -> AsyncGenerator[AsyncSession, None]:
+    async def get_async_db(cls) -> AsyncGenerator[AsyncSession, None]:
         """
         Get async database session
         """
         try:
             session_maker = cls._async_session_maker()
         except AttributeError:
-            cls.set_db()
+            cls.set_async_db()
             session_maker = cls._async_session_maker()
         async with session_maker as async_session:
             yield async_session
@@ -122,11 +133,11 @@ class DB:
         try:
             session_maker = cls._sync_session_maker()
         except AttributeError:
-            cls.set_db()
+            cls.set_sync_db()
             session_maker = cls._sync_session_maker()
         with session_maker as sync_session:
             yield sync_session
 
 
-get_db = DB.get_db
+get_db = DB.get_async_db
 get_sync_db = DB.get_sync_db